Move line-handler config to line module
